Input / Output

QuantumNeuralStates.save_master — Function

save_master(filename, θ, addrs; num_width=3)

Master function, that saves all important Neural Network configurations.

Example of txt file

Input Scaling:
    scale function: 
    max_norm:
    normalisation:

Weights:
    <Float32 values>

LayerNorm:
    true, false, ....

Inputs:
    0   0   0   0   0,   1   0   0   0   0, ...

QuantumNeuralStates.load_master — Function

load_master(chain, filename::String) -> x

Loads all configurations from plain text file into the Neural Network model chain and returns the saved inputs x as a matrix of Int (For reconstraction in Rimu notation). If only one column was saved, vec(x) is called to get a flat vector.

Arguments

chain: Neural Network model. See Chain.

filename: "nameoffile.txt".

QuantumNeuralStates.log_markov_chain — Function

log_markov_chain(filename, addrs; start=true, num_width)

Save one vector of addresses as one line in filename.

Keyword Arguments

start=true: create/overwrite file, write first line

start=false: append new line (no override)

num_width: character width reserved by integer (default 3)

Each address is written as its Int occupation-number vector, addresses separated by " | ".

Example

0, 1, 2, 3, 4, 5 | 1, 0, 0, 2, 1, 3 | ...

Index

QuantumNeuralStates.load_master

QuantumNeuralStates.log_markov_chain

QuantumNeuralStates.save_master