Training

QuantumNeuralStates.BlockStats — Type

BlockStats(energies, variance)

This structure acumulates each iteration's mean energy and variance. It is used in block_summary for statistics over block.

QuantumNeuralStates.block_summary — Function

block_summary(bs::BlockStats)
        -> E_mean, E_err, var_mean

This function calculates statistics of training data over blocks.

QuantumNeuralStates.TrainingPhase — Type

TrainingPhase

One stage of training. This structure allows to define more training phases with different parameters.

Variables

mode: define used loss function (and its gradients), see apply_loss! (default :energy).

optimiser: define what optimiser is used for parameters update (default :adam, others :minSR, :minSR_momentum, :descent). See also adam, descent, minSR.

vmc_sampler: define type of VMC sampler (default :metropolis). See also vmc_sample!.

stop: define a early-stop conditions, see StopBuffer.

η: learning rate used with parameters update (default 0.001f0).

λ: Tikhonov regularisation used in minSR optimiser (default 0.001f0).

skip: define how many VMC steps should be skip before parameters update is made. Thermalisation of VMC sampler (default (1, 100)). Multiple options can be stuck in Tuple of form (epoch, skip) - which means from epoch number epoch apply skip value skip. Example: [(1, 1000), (100, 500), (1000, 100)].

η_decrease: allows to decrease η learning rate and can be stuck in similar way as skip variable. (var_thr, factor) when variance threashold is reached than factor is applied to η as η_new = η*factor. Example: [(1, 0.1), (0.1, 0.1)] (defaultly not use).

truncation: if in NeuralAnsatz truncation is defined, this value allows to increase truncation with each phase. Allowing smoother increase in input truncation space. It is recommended to ONLY increase truncation value! (defaultly nothing).

block_size: define how many iterations should be stuck to block for printout (happens in blocks) and block statistics BlockStats, block_summary (default 10).

block_min: define how many blocks needs to be iterated before early stop-conditions can be applied (default 3).

patience: used in check_stop, if block-based stop condition is chosen, the condition is checked in last patience number of blocks (default 3).

max_epochs: maximum epochs in learning loop (default 1000). If none of early-stopping conditions are met this is safe-check stop condition.

Notes

QuantumNeuralStates.StopBuffer — Type

StopBuffer

This structure holds all possible options for early stopping in neural network training loop.

It is define either as Float64, value for chosen condition, or as Nothing, which means it is not used. By default all options are set to nothing.

Note

Special field is require_all, boolean. If true then all chosen conditions needs to be fulfilled before stop condition is triggered. If false than stop condition is triggered when any condition is fulfilled.

Example

stop = StopBuffer(var_thr=0.01)
stop = StopBuffer(var_thr=0.01, E_thr=0, require_all=true)

QuantumNeuralStates.check_stop — Function

check_stop(sb::StopBuffer, E_hist, var_hist, accept, patience)

This function is cheching if stop conditions chosen in StopBuffer have been fulfilled or not. If stop condition is defined over blocks, then it will look only in patience number of blocks for this condition.

QuantumNeuralStates.run_training_loop — Function

run_training_loop(H, ansatz, addr, phases; kwargs...)
        -> E_history, E_err_history, var_history, final_addrs

Run training through a sequence of TrainingPhase's. Each phase starts when the previous one converges (or hits max_epochs).

Variables

H: hamiltonian defined in Rimu.

ansatz: ansatz that evaluates wave-function. See NeuralAnsatz.

addr: Rimu stype of Fock address also used in H definition.

phases: details of each phase are saved in TrainingPhase and they can be stuck as: phases = [TrainingPhase(), TrainingPhase(), ...].

Keyword Arguments

savefile: name of file where parameters of NeuralAnsatz are saved. See also save_master. (default weights.txt)

loadfile: name of file from which parameters are loaded to NeuralAnsatz. See also load_master. (default weights.txt)

save: boolean value if I want to save parameters at the end of training loop (default true).

load: boolean value if I want to load saved parameters and weights for the training loop (default false).

markovfile: name of file where all Markov Chains of the training loop are saved. This is meant for diagnostics only as the number of Markov Chains can be enourmous (default MarkovChain.txt).

markov: boolean value if I want to save the Markov Chains (default false).

Index

QuantumNeuralStates.BlockStats

QuantumNeuralStates.StopBuffer

QuantumNeuralStates.TrainingPhase

QuantumNeuralStates.block_summary

QuantumNeuralStates.check_stop

QuantumNeuralStates.run_training_loop